# swagger.yaml
openapi: 3.0.0
info:
  title: XML-JSON Gateway API
  version: "2.1.0"
  description: >
    The XML-JSON Gateway API is a comprehensive service designed to facilitate data exchange 
    and conversion between XML and JSON formats. It provides endpoints to extract data from XML 
    using XPath, validate, transform, and format XML files, as well as convert XML to JSON and JSON to XML, 
    and verify/correct JSON files. All file paths are automatically mapped to internal storage.
    An API key is required in the request header ("x-api-key") for authorization.
    For an API access send an email to: bernhard@ikmail.com
servers:
  - url: http://127.0.0.1:3000
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
security:
  - ApiKeyAuth: []
...
paths:
  /extract:
    post:
      tags:
        - "XML Operations"
      summary: Extract Data from XML using XPath
      description: >
        Evaluates a provided XPath expression against an XML file. The API transforms the provided file path 
        to its internal storage location and returns the extracted value.
      requestBody:
        description: JSON payload containing the XML file path and the XPath expression.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filePath:
                  type: string
                  description: >
                    The file path of the XML file. The API maps this to internal storage automatically.
                  example: "/Volumes/Helmut/helmut480-test/A.xml"
                xpath:
                  type: string
                  description: The XPath expression to evaluate.
                  example: "count(//OM_RECORD)"
              required:
                - filePath
                - xpath
      responses:
        '200':
          description: Extraction successful.
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    description: The extracted value.
        '400':
          description: Invalid input, file not found, or wrong file extension.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Server encountered an error during extraction.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /validate:
    post:
      tags:
        - "XML Operations"
      summary: Validate XML File
      description: >
        Checks whether an XML file is well-formed. The API transforms the file path internally and returns a validation result.
      requestBody:
        description: JSON payload with the XML file path.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filePath:
                  type: string
                  description: >
                    The XML file path. The API maps this to internal storage automatically.
                  example: "/Volumes/Helmut/helmut480-test/A.xml"
              required:
                - filePath
      responses:
        '200':
          description: XML is valid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Invalid file path, file not found, or wrong file extension.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: XML validation error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /transform:
    post:
      tags:
        - "XML Operations"
      summary: Transform XML using XSLT
      description: >
        Applies an XSLT stylesheet to an XML file to produce transformed XML. The API transforms both file paths to internal storage and returns the transformed XML.
      requestBody:
        description: JSON payload with the source XML file path and the XSLT file path.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                xmlFilePath:
                  type: string
                  description: >
                    The source XML file path. Automatically mapped to internal storage.
                  example: "/Volumes/Helmut/helmut480-test/A.xml"
                xsltFilePath:
                  type: string
                  description: >
                    The XSLT file path. Automatically mapped to internal storage.
                  example: "/Volumes/Helmut/helmut480-test/style.xslt"
              required:
                - xmlFilePath
                - xsltFilePath
      responses:
        '200':
          description: Transformation successful.
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    description: The transformed XML.
        '400':
          description: Invalid file paths, file not found, or incorrect file extensions.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Transformation error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: string
  /format:
    post:
      tags:
        - "XML Operations"
      summary: Format XML File
      description: >
        Formats (pretty-prints) an XML file to improve readability. The API transforms the file path internally. 
        If a "saveTo" parameter is provided, the formatted XML is saved internally and a download URL is returned; 
        otherwise, the formatted XML is returned in the response.
      requestBody:
        description: JSON payload with the XML file path, an optional saveTo filename, and an optional logToConsole flag.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filePath:
                  type: string
                  description: >
                    The XML file path to format. Automatically mapped to internal storage.
                  example: "/Volumes/Helmut/helmut480-test/A.xml"
                saveTo:
                  type: string
                  description: >
                    Optional. Desired filename for the formatted XML. If provided, the file is saved internally and a download URL is returned.
                  example: "A_formatted.xml"
                logToConsole:
                  type: boolean
                  description: >
                    Optional. If true, prints the formatted XML to the server console.
                  example: false
              required:
                - filePath
      responses:
        '200':
          description: XML formatting successful.
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    description: >
                      The formatted XML (if "saveTo" is not provided).
                  message:
                    type: string
                    description: >
                      Confirmation message if the formatted XML was saved internally.
                  downloadUrl:
                    type: string
                    description: >
                      URL to download the formatted XML file (if "saveTo" is provided).
        '400':
          description: Invalid file path, file not found, or incorrect file extension.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Formatting error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: string
...
  /jsonverify:
    post:
      tags:
        - "JSON Operations"
      summary: Verify JSON File Validity
      description: >
        Reads a JSON file and checks whether it is valid by attempting to parse it. 
        The API transforms the file path internally and returns the validation result.
      requestBody:
        description: JSON payload with the JSON file path and an optional logToConsole flag.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filePath:
                  type: string
                  description: >
                    The JSON file path to verify. Automatically mapped to internal storage.
                  example: "/Volumes/Helmut/helmut480-test/A.json"
                logToConsole:
                  type: boolean
                  description: >
                    Optional. If true, additional logging is printed to the server console.
                  example: false
              required:
                - filePath
      responses:
        '200':
          description: JSON verification successful.
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    description: Indicates if the JSON is valid.
                  message:
                    type: string
                    description: A message indicating JSON validity.
        '400':
          description: Invalid file path, incorrect extension, or JSON file not found/invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  error:
                    type: string
                  details:
                    type: string
...